Set focus after creation of any new annotation.
