+++ app/assets/javascripts/lib/ui/content/annotations/replace.js
@@ -4,22 +4,9 @@ import throttle from 'lodash.throttle';
 import Component from 'lib/ui/component'
 import delegate from 'delegate';
 import debounce from 'debounce';
-import {editAnnotationHandle, stageChangeToAnnotation, stagePreviousContent, isEditable} from 'lib/ui/content/annotations';
+import {setFocus, editAnnotationHandle, stageChangeToAnnotation, stagePreviousContent, isEditable} from 'lib/ui/content/annotations';
 import {toggleElisionVisibility} from 'lib/ui/content/annotations/elide';
-import {getQueryStringDict} from 'lib/helpers';
 
-
-// Set focus to a particular replace on page load if specified in query string
-window.addEventListener('load', () => {
-  let query = getQueryStringDict();
-  if (query['annotation-id']) {
-    let annotation = document.querySelector(`.annotate.replacement[data-annotation-id="${query['annotation-id']}"]`);
-    if (annotation){
-      setFocus(annotation);
-      annotation.scrollIntoView(true);
-    };
-  }
-})
 
 // Respond to click, spacebar, or enter, like a real html button would
 delegate(document, '.annotate.replacement', 'click', e => handleReplaceButtonPressed(e));
@@ -55,13 +42,4 @@ function handleReplaceButtonPressed(e){
   }
 }
 
-function setFocus(el) {
-    if (document.activeElement === el) { return; }
-    var range = document.createRange();
-    var sel = window.getSelection();
-    range.setStart(el, 0);
-    range.collapse(true);
-    sel.removeAllRanges();
-    sel.addRange(range);
-    el.focus();
-}
+
